+++ client/rpc_request.go
@@ -16,29 +16,28 @@ import (
 
 type rpcRequest struct {
 	protocol    string
+	version     string
 	service     string
 	method      string
+	args        interface{}
 	contentType string
-	request     interface{}
 	opts        RequestOptions
 }
 
-func newRpcRequest(prootol, service, method string,
-	request interface{},
-	contentType string,
-	reqOpts ...RequestOption) Request {
+func newRpcRequest(protocol, version, service, method string, args interface{},
+	contentType string, reqOpts ...RequestOption) Request {
 
 	var opts RequestOptions
-
 	for _, o := range reqOpts {
 		o(&opts)
 	}
 
 	return &rpcRequest{
-		protocol:    prootol,
+		protocol:    protocol,
+		version:     version,
 		service:     service,
 		method:      method,
-		request:     request,
+		args:        args,
 		contentType: contentType,
 		opts:        opts,
 	}
@@ -48,6 +47,10 @@ func (r *rpcRequest) Protocol() string {
 	return r.protocol
 }
 
+func (r *rpcRequest) Version() string {
+	return r.version
+}
+
 func (r *rpcRequest) ContentType() string {
 	return r.contentType
 }
@@ -55,6 +58,7 @@ func (r *rpcRequest) ContentType() string {
 func (r *rpcRequest) ServiceConfig() registry.ServiceConfigIf {
 	return &registry.ServiceConfig{
 		Protocol: r.protocol,
+		Version:  r.version,
 		Service:  r.service,
 	}
 }
@@ -63,10 +67,14 @@ func (r *rpcRequest) Method() string {
 	return r.method
 }
 
-func (r *rpcRequest) Request() interface{} {
-	return r.request
+func (r *rpcRequest) Args() interface{} {
+	return r.args
 }
 
 func (r *rpcRequest) Stream() bool {
 	return r.opts.Stream
+}
+
+func (r *rpcRequest) Options() RequestOptions {
+	return r.opts
 }